Update tests, refactoring
